requirements.txt
tweepy
dune-client
python-dotenv
